+++ resources/js/actions/App/Http/Controllers/Household/ImportController.ts
@@ -1,51 +1,7 @@
 import { queryParams, type RouteQueryOptions, type RouteDefinition } from './../../../../../wayfinder'
 /**
-* @see \App\Http\Controllers\Household\ImportController::index
-* @see app/Http/Controllers/Household/ImportController.php:20
-* @route '/household/import'
-*/
-export const index = (options?: RouteQueryOptions): RouteDefinition<'get'> => ({
-    url: index.url(options),
-    method: 'get',
-})
-
-index.definition = {
-    methods: ["get","head"],
-    url: '/household/import',
-} satisfies RouteDefinition<["get","head"]>
-
-/**
-* @see \App\Http\Controllers\Household\ImportController::index
-* @see app/Http/Controllers/Household/ImportController.php:20
-* @route '/household/import'
-*/
-index.url = (options?: RouteQueryOptions) => {
-    return index.definition.url + queryParams(options)
-}
-
-/**
-* @see \App\Http\Controllers\Household\ImportController::index
-* @see app/Http/Controllers/Household/ImportController.php:20
-* @route '/household/import'
-*/
-index.get = (options?: RouteQueryOptions): RouteDefinition<'get'> => ({
-    url: index.url(options),
-    method: 'get',
-})
-
-/**
-* @see \App\Http\Controllers\Household\ImportController::index
-* @see app/Http/Controllers/Household/ImportController.php:20
-* @route '/household/import'
-*/
-index.head = (options?: RouteQueryOptions): RouteDefinition<'head'> => ({
-    url: index.url(options),
-    method: 'head',
-})
-
-/**
 * @see \App\Http\Controllers\Household\ImportController::start
-* @see app/Http/Controllers/Household/ImportController.php:27
+* @see app/Http/Controllers/Household/ImportController.php:24
 * @route '/household/import'
 */
 export const start = (options?: RouteQueryOptions): RouteDefinition<'post'> => ({
@@ -60,7 +16,7 @@ start.definition = {
 
 /**
 * @see \App\Http\Controllers\Household\ImportController::start
-* @see app/Http/Controllers/Household/ImportController.php:27
+* @see app/Http/Controllers/Household/ImportController.php:24
 * @route '/household/import'
 */
 start.url = (options?: RouteQueryOptions) => {
@@ -69,7 +25,7 @@ start.url = (options?: RouteQueryOptions) => {
 
 /**
 * @see \App\Http\Controllers\Household\ImportController::start
-* @see app/Http/Controllers/Household/ImportController.php:27
+* @see app/Http/Controllers/Household/ImportController.php:24
 * @route '/household/import'
 */
 start.post = (options?: RouteQueryOptions): RouteDefinition<'post'> => ({
@@ -77,6 +33,6 @@ start.post = (options?: RouteQueryOptions): RouteDefinition<'post'> => ({
     method: 'post',
 })
 
-const ImportController = { index, start }
+const ImportController = { start }
 
 export default ImportController