+++ resources/js/routes/household/import/index.ts
@@ -1,7 +1,7 @@
 import { queryParams, type RouteQueryOptions, type RouteDefinition } from './../../../wayfinder'
 /**
-* @see \App\Http\Controllers\Household\ImportController::index
-* @see app/Http/Controllers/Household/ImportController.php:20
+* @see \Illuminate\Routing\RedirectController::__invoke
+* @see vendor/laravel/framework/src/Illuminate/Routing/RedirectController.php:19
 * @route '/household/import'
 */
 export const index = (options?: RouteQueryOptions): RouteDefinition<'get'> => ({
@@ -10,13 +10,13 @@ export const index = (options?: RouteQueryOptions): RouteDefinition<'get'> => ({
 })
 
 index.definition = {
-    methods: ["get","head"],
+    methods: ["get","head","post","put","patch","delete","options"],
     url: '/household/import',
-} satisfies RouteDefinition<["get","head"]>
+} satisfies RouteDefinition<["get","head","post","put","patch","delete","options"]>
 
 /**
-* @see \App\Http\Controllers\Household\ImportController::index
-* @see app/Http/Controllers/Household/ImportController.php:20
+* @see \Illuminate\Routing\RedirectController::__invoke
+* @see vendor/laravel/framework/src/Illuminate/Routing/RedirectController.php:19
 * @route '/household/import'
 */
 index.url = (options?: RouteQueryOptions) => {
@@ -24,8 +24,8 @@ index.url = (options?: RouteQueryOptions) => {
 }
 
 /**
-* @see \App\Http\Controllers\Household\ImportController::index
-* @see app/Http/Controllers/Household/ImportController.php:20
+* @see \Illuminate\Routing\RedirectController::__invoke
+* @see vendor/laravel/framework/src/Illuminate/Routing/RedirectController.php:19
 * @route '/household/import'
 */
 index.get = (options?: RouteQueryOptions): RouteDefinition<'get'> => ({
@@ -34,8 +34,8 @@ index.get = (options?: RouteQueryOptions): RouteDefinition<'get'> => ({
 })
 
 /**
-* @see \App\Http\Controllers\Household\ImportController::index
-* @see app/Http/Controllers/Household/ImportController.php:20
+* @see \Illuminate\Routing\RedirectController::__invoke
+* @see vendor/laravel/framework/src/Illuminate/Routing/RedirectController.php:19
 * @route '/household/import'
 */
 index.head = (options?: RouteQueryOptions): RouteDefinition<'head'> => ({
@@ -44,8 +44,58 @@ index.head = (options?: RouteQueryOptions): RouteDefinition<'head'> => ({
 })
 
 /**
+* @see \Illuminate\Routing\RedirectController::__invoke
+* @see vendor/laravel/framework/src/Illuminate/Routing/RedirectController.php:19
+* @route '/household/import'
+*/
+index.post = (options?: RouteQueryOptions): RouteDefinition<'post'> => ({
+    url: index.url(options),
+    method: 'post',
+})
+
+/**
+* @see \Illuminate\Routing\RedirectController::__invoke
+* @see vendor/laravel/framework/src/Illuminate/Routing/RedirectController.php:19
+* @route '/household/import'
+*/
+index.put = (options?: RouteQueryOptions): RouteDefinition<'put'> => ({
+    url: index.url(options),
+    method: 'put',
+})
+
+/**
+* @see \Illuminate\Routing\RedirectController::__invoke
+* @see vendor/laravel/framework/src/Illuminate/Routing/RedirectController.php:19
+* @route '/household/import'
+*/
+index.patch = (options?: RouteQueryOptions): RouteDefinition<'patch'> => ({
+    url: index.url(options),
+    method: 'patch',
+})
+
+/**
+* @see \Illuminate\Routing\RedirectController::__invoke
+* @see vendor/laravel/framework/src/Illuminate/Routing/RedirectController.php:19
+* @route '/household/import'
+*/
+index.delete = (options?: RouteQueryOptions): RouteDefinition<'delete'> => ({
+    url: index.url(options),
+    method: 'delete',
+})
+
+/**
+* @see \Illuminate\Routing\RedirectController::__invoke
+* @see vendor/laravel/framework/src/Illuminate/Routing/RedirectController.php:19
+* @route '/household/import'
+*/
+index.options = (options?: RouteQueryOptions): RouteDefinition<'options'> => ({
+    url: index.url(options),
+    method: 'options',
+})
+
+/**
 * @see \App\Http\Controllers\Household\ImportController::start
-* @see app/Http/Controllers/Household/ImportController.php:27
+* @see app/Http/Controllers/Household/ImportController.php:24
 * @route '/household/import'
 */
 export const start = (options?: RouteQueryOptions): RouteDefinition<'post'> => ({
@@ -60,7 +110,7 @@ start.definition = {
 
 /**
 * @see \App\Http\Controllers\Household\ImportController::start
-* @see app/Http/Controllers/Household/ImportController.php:27
+* @see app/Http/Controllers/Household/ImportController.php:24
 * @route '/household/import'
 */
 start.url = (options?: RouteQueryOptions) => {
@@ -69,7 +119,7 @@ start.url = (options?: RouteQueryOptions) => {
 
 /**
 * @see \App\Http\Controllers\Household\ImportController::start
-* @see app/Http/Controllers/Household/ImportController.php:27
+* @see app/Http/Controllers/Household/ImportController.php:24
 * @route '/household/import'
 */
 start.post = (options?: RouteQueryOptions): RouteDefinition<'post'> => ({